rename setup-new-experiment ot setup-experimetn
