+++ pyproject.toml
@@ -20,9 +20,7 @@ dependencies = [
 
 [project.scripts]
 feb = "facts_experiment_builder.cli:main"
-facts-experiment-builder = "facts_experiment_builder.cli.cli:main"
 setup-new-experiment = "facts_experiment_builder.cli.setup_new_experiment_cli:main"
-setup-new-experiment-jinja2 = "facts_experiment_builder.cli.setup_new_experiment_jinja2_cli:main"
 generate-compose = "facts_experiment_builder.cli.generate_compose_cli:main"
 list-modules = "facts_experiment_builder.cli.list_modules_cli:list_modules"
